+++ src/main/java/com/gmail/netcracker/application/service/imp/FilterServiceImpl.java
@@ -19,7 +19,7 @@ public class FilterServiceImpl implements FilterService {
     private EventService eventService;
 
     @Override
-    public List<Event> filterOfPriority(List<Integer> validPriorities) {
+    public List<Event> filterOfPriority(List<Long> validPriorities) {
         List<Event> filterEvents = new ArrayList<>();
         for (Event event: eventService.myEventsWithPriority()){
             if(validPriorities.contains(event.getPriorityId()))
@@ -29,7 +29,7 @@ public class FilterServiceImpl implements FilterService {
     }
 
     @Override
-    public List<Event> filterOfType(List<Integer> validType) {
+    public List<Event> filterOfType(List<Long> validType) {
         List<Event> filterEvents = new ArrayList<>();
         for (Event event: eventService.getAllMyEvents()){
             if(validType.contains(event.getTypeId()))
